+++ sample/update_template.py
@@ -1,5 +1,5 @@
-from ecsapi_client import Api
+from ecsapi import Api
 
 api = Api()
 
-print(api.update_template(598, "update by ecsapi_client", "update by ecsapi_client"))
+print(api.update_template(598, "update by ecsapi", "update by ecsapi"))
