+++ src/ecsapi/_discount_record.py
