+++ src/ecsapi/_snapshot.py
