+++ src/ecsapi/errors.py
